+++ .gitignore
@@ -122,4 +122,13 @@ env/
 ## VS Code ###
 .vscode/
 
+## Selenium driver logs##
 geckodriver.log
+geckodriver.exe
+
+## vagrant files##
+.vagrant
+
+## Application Logs ##
+logs.txt
+applog.txt
