+++ pyproject.toml
@@ -9,11 +9,13 @@ python = "^3.9"
 flask = "^1.1.2"
 python-dotenv = "^0.14.0"
 requests = "^2.25.1"
-pytest = "^6.2.2"
+pytest = "^6.2.4"
 selenium = "^3.141.0"
 
-[tool.poetry.dev-dependencies]
 
+
+[tool.poetry.dev-dependencies]
+pylint = "^2.8.2"
 [build-system]
 requires = ["poetry>=0.12"]
 build-backend = "poetry.masonry.api"
